% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bootMean.R
\name{bootMean}
\alias{bootMean}
\title{bootmean function}
\usage{
bootMean(x, i)
}
\arguments{
\item{x}{the values}

\item{i}{a reference point}
}
\description{
This function calculates the mean of a vector and automatically removes NAs.
Used by getBoot.
}

