% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/bootSd.R
\name{bootSd}
\alias{bootSd}
\title{bootsd function}
\usage{
bootSd(x, i)
}
\arguments{
\item{x}{the values}

\item{i}{a reference point}
}
\description{
This function calculates the standard deviation of a vector and automatically
removes NAs. Used by getBoot.
}

