% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/doSpecies.R
\name{doSpecies}
\alias{doSpecies}
\title{Calculating Bootstrapped Means, SDs, and GLM Values By Species For Light
Calcs}
\usage{
doSpecies(trees, functionClass, functionRecord, response = "diayr",
  predictor = "light")
}
\arguments{
\item{trees}{The trees database, trimmed for light characters.}

\item{functionClass}{A parameter that takes either "boot" or "glm", to
indicate whether it should perform bootstrapping of the values or a glm
predicting response by predictor.}

\item{functionRecord}{A parameter that takes "mean" or "sd" if functionClass
is "boot"; it takes "slope" or "intercept" if functionClass is "glm"}

\item{response}{A string matching the response column's name in trees. Needed
for both "boot" and "glm" functions.}

\item{predictor}{A string matching the predictor column's name in trees.
Needed only for the "glm" function. Does not currently support multiple
predictors.}
}
\value{
This function returns a vector of calculated responses.
}
\description{
This is a function that takes a tree dataframe and performs calculations
based on funclass and funvar, on the column listed in "response"
}

