% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getBoot.R
\name{getBoot}
\alias{getBoot}
\title{getBoot function}
\usage{
getBoot(x, type = "mean", R = 1000)
}
\arguments{
\item{x}{A vector of values to resample.}

\item{type}{Default is "mean", set to "sd" to calculate standard deviation.}

\item{R}{The number of times to resample x.}
}
\value{
This function returns the mean or sd without any of the other
  characteristics returned by the boot() function.
}
\description{
This is a small function that bootstraps mean and sd from a given vector. It
requires the "boot" package to operate.
}
\examples{
getBoot(c(1:100), "mean", 1000)
}

