% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getCoefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{Calculating GLM Coefficients}
\usage{
getCoefficients(trees, functionRecord, response = "diayr",
  predictor = "light")
}
\arguments{
\item{trees}{The trees database, trimmed for light characters.}

\item{functionRecord}{A parameter that takes "slope" or "intercept" as its
possible return value.}

\item{response}{A string matching the response column's name in tree, or the
"y"}

\item{predictor}{A string matching the predictor column's name in trees, or
the "x". Needed only for the "glm"
function. Does not currently support multiple predictors.}
}
\value{
This function returns a vector of calculated responses.
}
\description{
This is a function that takes a tree dataframe and performs calculations
based on funclass and funvar, on the column listed in "response"
}
\examples{
dummydata <- data.frame(x=runif(50, 0, 2), y=runif(50, 0,25), SPCD=1)
getCoefficients(dummydata, "slope", response="y", predictor="x")
}

