% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getFIAchar.R
\name{getFIAchar}
\alias{getFIAchar}
\title{Generalized Function to Get the Mean of a Species Column (e.g., Height)}
\usage{
getFIAchar(sppinfo, treedb, char, percentage = TRUE)
}
\arguments{
\item{sppinfo}{A list of your species with spcd column}

\item{treedb}{Your reference tree database}

\item{char}{The character of interest's column name in treedb}

\item{percentage}{When TRUE, divides your results by 100.}
}
\value{
This function returns a mean for each species in sppinfo
}
\description{
Generalized Function to Get the Mean of a Species Column (e.g., Height)
}
\note{
I've used this for C2 and E1 in the sample code.
}

