% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getH2.R
\name{getH2}
\alias{getH2}
\title{Calculates the slope of asymptotic height for trees.}
\usage{
getH2(treedb, start = 1)
}
\arguments{
\item{treedb}{The tree dataframe, it must have H1, HT, and DIA columns.
It is generally assumed that the treedb you enter will be trimmed to
include only rows for a species or grouping.}

\item{start}{A starting number for the coefficient. 1, the default
value, is a pretty good guess for most.}
}
\description{
Calculates the slope of asymptotic height for trees.
}
\note{
This function calculates H2 from Pacala et al. (1996).
    It uses an nls function and is pretty specific and breakable,
    so please don't try anything too crazy with it.
}

