% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getMaxChar.R
\name{getMaxChar}
\alias{getMaxChar}
\title{Find the Max Value of a Tree Character}
\usage{
getMaxChar(codes, treedb, select = "HT")
}
\arguments{
\item{codes}{A list of tree codes (the numbers) that you
want to select for.}

\item{treedb}{The tree database you'll be working from.
It must have an SPCD column as well as the column in
select, default HT.}

\item{select}{Defaults to "HT", will accept any numeric
column in treedb.}
}
\description{
Find the Max Value of a Tree Character
}

