% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getStateData.R
\name{getStateData}
\alias{getStateData}
\title{Download state tree databases from the FIA Data Mart}
\usage{
getStateData(stateID, ...)
}
\arguments{
\item{stateID}{The two character state code for the data needed.}
}
\description{
A function to download state tree databases (ST\_TREE.CSV) from the FIA data mart, found at \url{http://apps.fs.fed.us/fiadb-downloads/datamart.html}. This function may take a long time, because the data are large and the FIA Data Mart may be slow. You will receive an immediate error if the data are not available. Header=T is the only paramater passed into the read.csv function, you can pass others after the stateID parameter.
}

