% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/prepLightData.R
\name{prepLightData}
\alias{prepLightData}
\title{Trimming a Data Frame to Entries with Light and Age Characters}
\usage{
prepLightData(treedb)
}
\arguments{
\item{treedb}{The TreesCA dataframe or similar. No customization options
available.}
}
\value{
This function returns a dataframe with light and diayr response columns.
}
\description{
This is a cheap little function I made that trims TreesCA or similar to only
entries with both CLIGHTCD and BHAGE, and converts CLIGHTCD into a more
appropate 0-100 style percentage of light. Finally, it calculates the
diameter increase per year for each tree by dividing DIA by BHAGE. It returns
a dataframe with light and diayr columns.
}
\examples{
store <- prepLightData(TreesCA)
head(store)
}

