% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pullSpeciesCodes.R
\name{pullSpeciesCodes}
\alias{pullSpeciesCodes}
\title{Get Species Codes From A Reference Table}
\usage{
pullSpeciesCodes(speciesList, refDataFrame = "")
}
\arguments{
\item{speciesList}{A character vector of the species shortcodes
(look them up in USDA PLANTs database)}

\item{refDataFrame}{A dataframe containing at least two columns:
SPECIES_SYMBOL and SPCD}
}
\value{
Returns a vector of species codes from the reference dataframe.
}
\description{
Get Species Codes From A Reference Table
}
\examples{
sampleList <- c("ABCO", "ABMA")
## RefSpecies is an included reference species list for CA

pullSpeciesCodes(sampleList, RefSpecies)
}

