% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pullTrees.R
\name{pullTrees}
\alias{pullTrees}
\title{Get Individual Tree Records From A Reference FIA datafile}
\usage{
pullTrees(sppcodes, treedb, select = c("PLT_CN", "TREE", "SPCD", "DIA", "HT",
  "CR", "CDENCD", "TRANSCD", "BHAGE", "TOTAGE", "CLIGHTCD"))
}
\arguments{
\item{sppcodes}{A vector of the species spcd numbers. These are
found in the FIA's REF_SPECIES.csv file for your state, or can be
imported using \code{\link{pullSpeciesCodes}} function.}

\item{treedb}{A dataframe containing the columns specified
in select, along with an SPCD column.}

\item{select}{The columns that you want to keep from the larger treedb.}
}
\value{
Returns a subsetted version of treedb with just the species you requested.
}
\description{
Get Individual Tree Records From A Reference FIA datafile
}
\examples{
trees <- read.csv("http://apps.fs.fed.us/fiadb-downloads/CA_TREE.CSV", stringsAsFactors=FALSE)
sampleSppCodes <- c(15, 20, 81)
store <- pullTrees(sampleSppCodes, trees)
}

