% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/putH2.R
\name{putH2}
\alias{putH2}
\title{Returns a Vector of Slope of Asymptotic Height Values For Species}
\usage{
putH2(sppinfo, treedb)
}
\arguments{
\item{sppinfo}{Should be a dataframe with "spcd" column. Need to update,
because it could easily be a vector.}

\item{treedb}{Your reference tree database, as full as needed.}
}
\description{
This function is a wrapper for \code{\link{getH2}}, to sort by species
  instead of assuming that you already put the trimmed treedb file in or only
  need one go of it. Read \code{\link{getH2}} for more information about how this parameter is calculated.
}

