% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/putMaxHeight.R
\name{putMaxHeight}
\alias{putMaxHeight}
\title{Put Maximum Heights Into TreeDB According To SPCD}
\usage{
putMaxHeight(sppinfo, treespcd)
}
\arguments{
\item{sppinfo}{Must be a dataframe with spcd and maxht columns.}

\item{treespcd}{Must be a vector or dataframe with SPCD column.}
}
\value{
Returns a vector matching the length of treespcd with
    maximum heights in appropriate places.
}
\description{
Put Maximum Heights Into TreeDB According To SPCD
}

