% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/readCSVs.R
\name{readCSVs}
\alias{readCSVs}
\title{Read CSV files from a directory}
\usage{
readCSVs(directory = "", lower = TRUE)
}
\arguments{
\item{directory}{A character vector of the file path, leave blank to search in your current working directory.}

\item{lower}{If set to true, all variable will be lower case.}
}
\description{
Read CSV files from a directory
}
\examples{
readCSVs("", TRUE)
}
\keyword{csv}
\keyword{import}

