% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate Standard Error From A Vector}
\usage{
se(x)
}
\arguments{
\item{x}{A single number or vector of numbers}
}
\value{
Returns the standard error of x. Automatically removes NA.
}
\description{
Calculate Standard Error From A Vector
}
\examples{
se(c(5,10,12,13))
}

