% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/TreesCA.R
\docType{data}
\name{TreesCA}
\alias{TreesCA}
\title{TreesCA Dataset}
\format{A dataframe with 181,891 observations in 11 columns
    of individual trees records from CA.
\describe{
   \item{PLT_CN}{The plot identifier.}
   \item{TREE}{The tree identifier.}
   \item{SPCD}{The four digit species code, associated with
       RefSpecies.}
   \item{DIA}{Diameter at base height of trees, in inches.}
   \item{HT}{The height of a tree in feet, including
       estimation if the tree was broken from wind damage.}
   \item{CR}{The compacted crown ratio of a tree. This
       excludes open areas within the crown.}
   \item{CDENCD}{The crown density of a tree.}
   \item{TRANSCD}{The transparency of a crown of a tree.}
   \item{BHAGE}{The breast height age of a tree, estimated from
       a tree core taken at breast height.}
   \item{TOTAGE}{The total age of a tree, estimated from a
       tree core.}
   \item{CLIGHTCD}{The amount of light that a tree receives. }

}}
\usage{
TreesCA
}
\description{
This dataframe is an example of what "pullSpeciesCodes" and
 "pullTrees" can do. This is our working dataset for most
 of our analyses, and has trimmed the full ca_tree CSV file
 down from 26.1 MB to 1.6MB, a much more manageable file.
 This file and its predecessor (FullTreesCA) are from the
 FIA Datamart at
 \url{http://apps.fs.fed.us/fiadb-downloads/datamart.html}.
 Please refer to the documentation for further details about
 these data:
 \url{http://www.fia.fs.fed.us/library/database-documentation/}
}
\keyword{datasets}

