% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/getRadius.R
\name{getRadius}
\alias{getRadius}
\title{Get the Crown Radius Of Individual Trees}
\usage{
getRadius(speq, treedb)
}
\arguments{
\item{speq}{This is a dataframe with the species codes (numbers)
and b0 and b1 for the linear regressions.}

\item{treedb}{This is your reference tree database.}
}
\value{
This function returns a list of crown radii
    matching the treedb length for your species.
}
\description{
Get the Crown Radius Of Individual Trees
}
\note{
This does a linear regression to calculate crown radius
       from diameter (DIA) using species-specific intercept and
       slope.
}

