% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/putH2.R
\name{putH2}
\alias{putH2}
\title{Returns a Vector of H2 Values For Species}
\usage{
putH2(sppinfo, treedb)
}
\arguments{
\item{sppinfo}{Should be a dataframe with "spcd" column. Need to update, because it could easily be a vector.}

\item{treedb}{Your reference tree database, as full as needed.}
}
\description{
Returns a Vector of H2 Values For Species
}
\note{
This function is basically a wrapper for getH2, to sort
    by species instead of assuming that you already put the
    trimmed treedb file in or only need one go of it.
}

