% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/unitConvert.R
\name{unitConvert}
\alias{unitConvert}
\title{Convert To And From Standard And Metric Measurements}
\usage{
unitConvert(x, invar, outvar)
}
\arguments{
\item{x}{A single number or vector of numbers to convert}

\item{invar}{The units of x. Acceptable inputs are: ft, in, cm, m, in3.}

<<<<<<< HEAD
\item{outvar}{The desired units of x. Acceptable outputs are: ft, in, cm, m, cm3.
cm3 is only acceptable if the inputvar is in3.}
=======
\item{outvar}{The desired units of x. Acceptable outputs are: ft, in, cm, m, cm3.}
>>>>>>> 6b627a8d487106cd02b21b97030915d64d32de70
}
\value{
Returns x, transformed into the appropriate units.
}
\description{
Convert To And From Standard And Metric Measurements
}
\examples{
unitConvert(2, "in", "cm")
}

